Rename component folder: `Withdrawal` => `TransferService`
