+++ src/pages/account.tsx
@@ -15,7 +15,7 @@ import withFallback from "../components/Lazy/withFallback"
 import PaymentDialog from "../components/Payment/PaymentDialog"
 import ReceivePaymentDialog from "../components/Payment/ReceivePaymentDialog"
 import ViewLoading from "../components/ViewLoading"
-import WithdrawalDialog from "../components/Withdrawal/WithdrawalDialog"
+import WithdrawalDialog from "../components/TransferService/WithdrawalDialog"
 import { Account, AccountsContext } from "../context/accounts"
 import { useLiveAccountData } from "../hooks/stellar-subscriptions"
 import { useIsMobile, useRouter } from "../hooks/userinterface"
