+++ stories/Withdrawal.tsx
@@ -3,10 +3,10 @@ import { Asset } from "stellar-sdk"
 import { WithdrawalSuccessResponse } from "@satoshipay/stellar-sep-6"
 import { action } from "@storybook/addon-actions"
 import { storiesOf } from "@storybook/react"
-import WithdrawalTransactionForm from "../src/components/Withdrawal/WithdrawalTransactionForm"
-import AnchorWithdrawalInitForm from "../src/components/Withdrawal/WithdrawalRequestForm"
-import WithdrawalKYCRedirect from "../src/components/Withdrawal/WithdrawalKYCRedirect"
-import WithdrawalKYCStatus from "../src/components/Withdrawal/WithdrawalKYCStatus"
+import WithdrawalTransactionForm from "../src/components/TransferService/WithdrawalTransactionForm"
+import AnchorWithdrawalInitForm from "../src/components/TransferService/WithdrawalRequestForm"
+import WithdrawalKYCRedirect from "../src/components/TransferService/WithdrawalKYCRedirect"
+import WithdrawalKYCStatus from "../src/components/TransferService/WithdrawalKYCStatus"
 import { Account } from "../src/context/accounts"
 import { RefStateObject } from "../src/hooks/userinterface"
